App theme switcher architecture some refactor
