+++ app/src/main/java/com/example/yp_playlist_maker/App.kt
@@ -31,7 +31,7 @@ class App : Application() {
 
         val getThemeColor = Creator.provideAppThemeInteractor(this)
 
-        darkTheme = getThemeColor.getSwitcherStatus()
+        darkTheme = getThemeColor.getAppTheme()
         switchTheme(darkTheme)
 
     }
