+++ app/src/main/java/com/example/yp_playlist_maker/Creator.kt
@@ -1,15 +1,16 @@
 package com.example.yp_playlist_maker
 
 import android.content.Context
-import com.example.yp_playlist_maker.data.AppThemeImpl
-import com.example.yp_playlist_maker.domain.api.AppThemeInteractor
-import com.example.yp_playlist_maker.domain.api.AppThemeRepository
-import com.example.yp_playlist_maker.domain.impl.AppThemeInteractorImpl
+import com.example.yp_playlist_maker.data.impl.app_theme.AppThemeImpl
+import com.example.yp_playlist_maker.data.storage.app_theme.SharedPrefsAppThemeStorage
+import com.example.yp_playlist_maker.domain.api.app_theme.AppThemeInteractor
+import com.example.yp_playlist_maker.domain.api.app_theme.AppThemeRepository
+import com.example.yp_playlist_maker.domain.impl.app_theme.AppThemeInteractorImpl
 
 object Creator {
 
     private fun getAppThemeRepository(context: Context): AppThemeRepository {
-        return AppThemeImpl(context)
+        return AppThemeImpl(SharedPrefsAppThemeStorage(context))
     }
 
     fun provideAppThemeInteractor(context: Context): AppThemeInteractor {
