--- app/src/main/java/com/example/yp_playlist_maker/domain/api/AppThemeRepository.kt
@@ -1,11 +1,0 @@
-package com.example.yp_playlist_maker.domain.api
-
-import com.example.yp_playlist_maker.domain.models.AppThemeParams
-
-interface AppThemeRepository {
-
-    fun saveSwitcherStatus(params: AppThemeParams)
-
-    fun getSwitcherStatus(): Boolean
-
-}
