+++ app/src/main/java/com/example/yp_playlist_maker/domain/models/app_theme/AppThemeParams.kt
@@ -1,4 +1,4 @@
-package com.example.yp_playlist_maker.domain.models
+package com.example.yp_playlist_maker.domain.models.app_theme
 
 data class AppThemeParams (
     val switcherStatus: Boolean
